Gui: split selection visibility implementation
