+++ src/Gui/Selection/Selection.h
